does some small cleanup tasks to keep future pr focused on refactor.  reverts some accidentaly dropped behaviors. removes unused layout field and makes internal subpanels package-private, reverts some accidental changes in .project
